+++ meinberlin/apps/budgeting/templates/meinberlin_budgeting/proposal_moderate_form.html
@@ -28,9 +28,9 @@
                 {{ form.media }}
             {% endfor %}
 
-            {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.proposal.moderator_feedback %}
+            {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.moderateable.moderator_feedback %}
             {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.statement.statement %}
-            {% include 'meinberlin_contrib/includes/form_checkbox_field.html' with field=forms.proposal.is_archived %}
+            {% include 'meinberlin_contrib/includes/form_checkbox_field.html' with field=forms.moderateable.is_archived %}
 
             <div class="u-spacer-bottom">
                 <input type="submit" class="btn btn--primary" value="{% trans 'Save' %}" />
