+++ meinberlin/apps/budgeting/views.py
@@ -1,17 +1,11 @@
 import django_filters
-from django.db import transaction
 from django.utils.translation import ugettext_lazy as _
-from django.views import generic
 
 from adhocracy4.categories import filters as category_filters
 from adhocracy4.filters import filters as a4_filters
-from adhocracy4.projects.mixins import ProjectMixin
-from adhocracy4.rules import mixins as rules_mixins
-from meinberlin.apps.contrib import forms as contrib_forms
 from meinberlin.apps.contrib import filters
 from meinberlin.apps.ideas import views as idea_views
-from meinberlin.apps.moderatorfeedback.forms import ModeratorStatementForm
-from meinberlin.apps.moderatorfeedback.models import ModeratorStatement
+from meinberlin.apps.ideas.views import AbstractIdeaModerateView
 from meinberlin.apps.projects.views import ArchivedWidget
 
 from . import forms
@@ -90,50 +84,8 @@ class ProposalDeleteView(idea_views.AbstractIdeaDeleteView):
     template_name = 'meinberlin_budgeting/proposal_confirm_delete.html'
 
 
-class ProposalModerateView(ProjectMixin,
-                           rules_mixins.PermissionRequiredMixin,
-                           generic.detail.SingleObjectMixin,
-                           generic.detail.SingleObjectTemplateResponseMixin,
-                           contrib_forms.BaseMultiModelFormView):
+class ProposalModerateView(AbstractIdeaModerateView):
     model = models.Proposal
     permission_required = 'meinberlin_budgeting.moderate_proposal'
     template_name = 'meinberlin_budgeting/proposal_moderate_form.html'
-    get_context_from_object = True
-
-    forms = {
-        'proposal': {
-            'model': models.Proposal,
-            'form_class': forms.ProposalModerateForm
-        },
-        'statement': {
-            'model': ModeratorStatement,
-            'form_class': ModeratorStatementForm
-        }
-    }
-
-    def dispatch(self, *args, **kwargs):
-        self.object = self.get_object()
-        return super().dispatch(*args, **kwargs)
-
-    def get_success_url(self):
-        return self.object.get_absolute_url()
-
-    def forms_save(self, forms, commit=True):
-        objects = super().forms_save(forms, commit=False)
-        proposal = objects['proposal']
-        statement = objects['statement']
-
-        if not statement.pk:
-            statement.creator = self.request.user
-
-        with transaction.atomic():
-            statement.save()
-            proposal.moderator_statement = statement
-            proposal.save()
-        return objects
-
-    def get_instance(self, name):
-        if name == 'proposal':
-            return self.object
-        elif name == 'statement':
-            return self.object.moderator_statement
+    moderateable_form_class = forms.ProposalModerateForm
