+++ meinberlin/apps/contrib/templates/meinberlin_contrib/item_detail.html
@@ -24,6 +24,11 @@
                     <div class="label label--big">{{ object.category }}</div>
                 {% endif %}
                 {% block additional_labels %}{% endblock %}
+                {% if object.moderator_feedback %}
+                    <span class="label label--big label--{{ object.moderator_feedback|classify }}">
+                        {{ object.get_moderator_feedback_display }}
+                    </span>
+                {% endif %}
             </div>
 
             <div class="item-detail__content">
@@ -77,6 +82,16 @@
                                     <a class="dropdown-item" href="{{ delete_url }}">{% trans 'Delete' %}</a>
                                 </li>
                             {% endif %}
+
+                            {% get_item_permission object 'moderate' as moderate_perm %}
+                            {% has_perm moderate_perm request.user object as is_moderator %}
+                            {% get_item_url object 'moderate' False as moderate_url %}
+                            {% if is_moderator and moderate_url %}
+                                <li>
+                                    <a class="dropdown-item" href="{{ moderate_url }}">{% trans 'Moderate' %}</a>
+                                </li>
+                            {% endif %}
+
                             {% block dropdown_items %}{% endblock %}
                             <li>
                                 {% trans 'Report' as report_text %}
@@ -89,7 +104,26 @@
         </div>
     </div>
 
-    {% block moderator_statement %}{% endblock %}
+    {% block moderator_statement %}
+        {% if object.moderator_statement.statement %}
+            <div class="l-wrapper">
+                <div class="l-center-8">
+                    <div class="moderator-statement">
+                        <h2 class="moderator-statement__title">{% trans 'Official Statement' %}</h2>
+                        <div class="moderator-statement__subtitle">
+                            <strong class="moderator-statement__organisation">
+                                {{ object.module.project.organisation.name }}
+                            </strong>
+                            {% html_date object.moderator_statement.created class='moderator-statement__date' %}
+                        </div>
+                        <div class="moderator-statement__body">
+                            {{ object.moderator_statement.statement | safe }}
+                        </div>
+                    </div>
+                </div>
+            </div>
+        {% endif %}
+    {% endblock %}
 
     <div class="l-center-6">
         {% react_comments object %}
