+++ meinberlin/apps/contrib/templatetags/item_tags.py
@@ -1,5 +1,6 @@
 from django import template
 from django.core.urlresolvers import reverse
+from django.urls import NoReverseMatch
 
 register = template.Library()
 
@@ -44,14 +45,19 @@ def get_item_delete_url(item):
 
 
 @register.assignment_tag
-def get_item_url(item, view):
+def get_item_url(item, view, raises=True):
     url_name = '{app}:{name}-{view}'.format(
         app=item._meta.app_label,
         name=item._meta.model_name,
         view=view
     )
 
-    return reverse(url_name, kwargs={
-        'year': item.created.year,
-        'pk': '{:05d}'.format(item.pk)
-    })
+    try:
+        return reverse(url_name, kwargs={
+            'year': item.created.year,
+            'pk': '{:05d}'.format(item.pk)
+        })
+    except NoReverseMatch:
+        if raises:
+            raise
+        return ''
