+++ meinberlin/apps/kiezkasse/templates/meinberlin_kiezkasse/proposal_moderate_form.html
@@ -27,7 +27,7 @@
                 {{ form.media }}
             {% endfor %}
 
-            {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.proposal.moderator_feedback %}
+            {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.moderateable.moderator_feedback %}
             {% include 'meinberlin_contrib/includes/form_field.html' with field=forms.statement.statement %}
